test: move URL scheme validation test to tests directory
Addresses Greptile review feedback: embedded test was not discoverable
by pytest. Created separate tests/test_browser_lynx.py for lynx backend
tests, avoiding the playwright skip in test_browser.py.
