+++ gptme/tools/_browser_lynx.py
@@ -69,24 +69,3 @@ def test_search():
     # result = search("Python", "google")
     result = search("Erik Bjäreholt", "duckduckgo")
     assert "erik.bjareholt.com" in result
-
-
-def test_url_scheme_validation():
-    """Test that dangerous URL schemes are blocked."""
-    import pytest
-
-    # Valid schemes should work (will fail if lynx not installed, but that's OK)
-    # We're testing the validation, not the actual fetch
-    _validate_url_scheme("https://example.com")
-    _validate_url_scheme("http://example.com")
-    _validate_url_scheme("HTTP://EXAMPLE.COM")  # Case insensitive
-
-    # Dangerous schemes should be blocked
-    with pytest.raises(ValueError, match="not allowed"):
-        _validate_url_scheme("file:///etc/passwd")
-
-    with pytest.raises(ValueError, match="not allowed"):
-        _validate_url_scheme("ftp://example.com")
-
-    with pytest.raises(ValueError, match="not allowed"):
-        _validate_url_scheme("javascript:alert(1)")
